requirements_test.txt
coverage==4.2
mock>=1.0.1
flake8>=2.1.0
tox>=1.7.0


# Additional test requirements go here
django
django-mptt
Unidecode